sse/requirements.txt
fastapi==0.115.5
openai==1.55.1
psycopg2==2.9.10
pydantic==2.10.2
python-dotenv==1.0.1
